scrape/requirements.txt
Pillow==6.1.0
requests==2.22.0
urllib3==1.25.3
PyDrive==1.3.1
loguru==0.3.0
piexif==1.1.3
